[refactor] import 구문 수정
