+++ client/index.js
@@ -3,7 +3,7 @@ import { AppRegistry } from 'react-native';
 import { App } from './App';
 import { name as appName } from './app.json';
 import { Provider } from 'react-redux';
-import globalStore from './src/Contexts/Common//globalStore';
+import globalStore from './src/Contexts/Common/globalStore';
 const store = globalStore();
 
 const RNRedux = () => (
