+++ client/src/Contexts/Common/globalStore.js
@@ -1,5 +1,5 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
-import userReducer from '../User/reducer';
+import userReducer from '../User';
 import ReduxThunk from 'redux-thunk';
 const rootReducer = combineReducers({ user: userReducer });
 const globalStore = () => {
